Add files via upload
Tricks learned from this script: When do data frame transpose, the important thing to remember is that we should make sure the cells (or each element in the df) are numeric values, usually, in data frame, we also have some other variables as factors or characters, remove it or transform it to rownames, or keep it separately. In our case here, we transform the character column as rownames, and then remove that column, after this procedure, all cells are numeric, and then transform the df to matrix by as.matrix, and then transpose it by t(),if we want to transform matrix back to df, we can do as.data.frame(), and now the previous rownames of old df becomes the colnames of new df.